Split Lazaret Controller into 2 classes
